refactor: rewrite fetch module
BREAKING CHANGE: Removed fetchJson; refactored fetch to accept url as a separate parameter, matching the web standard API.
